chore(cleanup): remove temporary file
